Progress on the FROM, grouping, and aggregate
In the test file, a change was made to reflect what was
said in Piazza. In the DataPage file, a small bug was fixed
to fix that we also need to allocate +2 to offset for the slot.
The FROM stuff is mostly taken care of I believe. The grouping
and aggregation is still in progress.