Completed functionality of ANALYZE
First, the addTuple() method was reverted to the original version
to take care of bugs. Note that also the getTupleDataEnd() method
was reverted back to the Donnie's original code. The analyze()
method was completed following what was asked for in the problem set.
This was tested manually and seemed to work for simple cases.