Changes in CostBasedJoinPlanner.makePlan()
I attempted to fill in some logic for the makePlan() but
it is probably wrong so I will go to OH to figure it out.
Specifically, I tried to utilize makeOptimalJoinPlan
whenever there is a join expression in the fromClause, and
if not, I do the usual completeFromClause(). This is the plan,
not sure if it is correct yet.