Cleaned up CostBaseJoinPlanner.java
Made changes in makeLeafNode(), where there was one line
that was causing many errors to fail, but that has been
taken care of. Also made changes in makePlan(). I am testing
the code by running ant test and checking what tests we fail
in the index.html file from Assignment 2. Now, most tests are
passing, but 4 are failing that we have been passing before.
This might be because of an ordering issue with columns that might
not matter that much based on a Piazza post.