Fixed BTreeTupleFile.java to pass 7/8 tests
I added a break statement in BTreeTupleFile.java, which fixes 3
more tests. What was happening was that the nextPageNo kept
updating even after it satisfied the if statement once because
keys in a page are increasing in order. I made a change in
PageTuple for debugging purposes and basically found out that
tuple in that file was null which was causing the
nullPointerException. First, I thought this might be because
we aren't setting the newLeaf's pageNo, but I believe this
is done when it is initialized. There might be a problem in
the code we wrote for the third part.